Split poll types
