+++ packages/types/src/index.ts
@@ -12,6 +12,7 @@ export * from './discord/guildTemplates.js'
 export * from './discord/interactions.js'
 export * from './discord/invites.js'
 export * from './discord/messages.js'
+export * from './discord/polls.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
